Make AddCommand and subclasses consistent and minimal
* Implement equality, hashcode, toString in AddCommand<T>
* Abstract out duplicate messages into respective class
