+++ src/main/java/seedu/address/logic/ContactMessages.java
@@ -1,0 +1,10 @@
+package seedu.address.logic;
+
+/**
+ * Container for user visible messages related to {@code Contact}.
+ */
+public class ContactMessages {
+
+    public static final String MESSAGE_DUPLICATE_CONTACT = "This contact already exists in the address book";
+
+}
