+++ src/main/java/seedu/address/logic/commands/create/AddTodoCommand.java
@@ -5,29 +5,29 @@ import static seedu.address.logic.parser.CliSyntax.TODO_COMMAND_WORD;
 import static seedu.address.logic.parser.todo.TodoCliSyntax.PREFIX_TODO_DEADLINE_LONG;
 import static seedu.address.logic.parser.todo.TodoCliSyntax.PREFIX_TODO_LOCATION_LONG;
 import static seedu.address.logic.parser.todo.TodoCliSyntax.PREFIX_TODO_NAME_LONG;
+import static seedu.address.logic.parser.todo.TodoCliSyntax.PREFIX_TODO_TAG_LONG;
 
-import seedu.address.commons.util.ToStringBuilder;
 import seedu.address.logic.Messages;
 import seedu.address.model.Model;
 import seedu.address.model.todo.Todo;
 
 /**
- * Adds a todo.
+ * Adds a {@code Todo} to the app.
  */
 public class AddTodoCommand extends AddCommand<Todo> {
-
-    public static final String COMMAND_WORD = "add";
 
     public static final String MESSAGE_USAGE = TODO_COMMAND_WORD + " " + COMMAND_WORD
             + ": Adds a todo to the app. "
             + "Parameters: "
             + PREFIX_TODO_NAME_LONG + " NAME "
-            + PREFIX_TODO_DEADLINE_LONG + " PREFIX_TODO_DEADLINE "
-            + PREFIX_TODO_LOCATION_LONG + " LOCATION\n"
+            + PREFIX_TODO_DEADLINE_LONG + " DEADLINE "
+            + PREFIX_TODO_LOCATION_LONG + " LOCATION "
+            + "[" + PREFIX_TODO_TAG_LONG + " TAG]...\n"
             + "Example: " + TODO_COMMAND_WORD + " " + COMMAND_WORD + " "
             + PREFIX_TODO_NAME_LONG + " Grading students projects "
             + PREFIX_TODO_DEADLINE_LONG + " 25-05-23 17:00 "
-            + PREFIX_TODO_LOCATION_LONG + " NUS Science ";
+            + PREFIX_TODO_LOCATION_LONG + " NUS Science "
+            + PREFIX_TODO_TAG_LONG + " important ";
 
     public static final String MESSAGE_SUCCESS = "New todo added: %1$s";
 
@@ -48,24 +48,4 @@ public class AddTodoCommand extends AddCommand<Todo> {
         return String.format(MESSAGE_SUCCESS, Messages.format(todo));
     }
 
-    @Override
-    public boolean equals(Object other) {
-        if (other == this) {
-            return true;
-        }
-
-        // instanceof handles nulls
-        if (!(other instanceof AddTodoCommand otherAddTodoCommand)) {
-            return false;
-        }
-
-        return itemToAdd.equals(otherAddTodoCommand.itemToAdd);
-    }
-
-    @Override
-    public String toString() {
-        return new ToStringBuilder(this)
-                .add("toAdd", itemToAdd)
-                .toString();
-    }
 }
